(*)Use thickness variables in hybgenaij_regrid
  Revised hybgenaij_regrid to work with thickness variables along side the
interface height variables to reduce the impacts of roundoff and to be able to
handle layers that are very thin compared with the depth of the ocean.  Also
moved cushn into a separate function outside of hybgenaij_regrid.  All
expressions are mathematically equivalent, and there are could be changes in
solutions that use hygben remapping, but answers are bitwise identical in all
other cases.