(*)Avoid recalculating thicknesses in hybgen_remap
  The old and new grid thicknesses are now arguments to hygben_remap, but in the
original code imported from HYCOM it was the interface heights.  So to preserve
the old answers the interface heights were being constructed from the
thicknesses and then the thicknesses are being reconstructed from the interface
heights, which leads to differences at the level of roundoff.  This commit
eliminates this unnecessary reconstruction step, which changes answers in cases
using Hybgen remapping at the level of roundoff, but all other cases are bitwise
identical.