(*)Bug fix in hybgen_column_regrid
  Corrected the calculation of the layer thickness and interface height changes
when entraining into the uppermost fixed layer, correcting a bug that was
introduced during code reformulations, and was not present in Hycom.  Before
this update, the model was failing a self-consistency test that was recently
added at the end of hybgen_column_regrid.  This does change (correct) answers in
cases using Hybgen regridding, but answers in other cases are bitwise identical.