(*)Set dpu from the average of dp in hybgen_remap
  Set the thickness at velocity points with Hybgen remapping directly from the
average of the adjacent tracer points thicknesses and then filter out the
bottom, rather than setting them based on the difference in the interface
positions averaged to velocity points, so that away from the bottom, no
subtraction of large numbers occurs.  This is mathematically equivalent but
changes answers at roundoff when USE_HYBGEN_REMAP is true.  All answers are
bitwise identical in other cases.