(*)Reuse h_tot in partial cell masks for remapping
  Use h_tot based on the old thicknesses when applying the partial cell masking
to the thicknesses at velocity points that are used to remap velocities with
Hybgen remapping.  The total thicknesses are conserved during remapping, but
their sums are mathematically equivalent but differ at roundoff, so this does
change answers at roundoff.  However, this change does avoid an extra vertical
sum of a 3-d array.  Answers are bitwise identical if Hybgen remapping is not
used for velocities.