+Removed Hybgen thbase variables
  Removed the various thbase variables that the Hycom hybgen code uses to
introduce and arbitrary offset for the potential density variables that the
model carries around, following advice from Alan Wallcraft.  Now, like the rest
of the MOM6 code, all density variables used by tht MOM6 Hybgen code are simply
the full densities.  This capability to use an arbitrary was not being exercised
yet, so no answers change, but it would impact any cases that exercised the
older versions of the hybgen_unmix code.