(*)Eliminated HYBGEN_N_HYBRID
  Eliminated the ability set the number of hybrid levels with hybgen regridding.
All cases now operate as though N_HYBRID had be set to NK, which is how Hybgen
was almost always used in Hycom.  Several subroutines or other blocks of code
that are no longer used after this change were also eliminated. This does change
answers if Hybgen regridding is used and HYBGEN_N_HYBRID was not set to NK, but
answers in other cases are bitwise identical.