(*)Split hybgen regrid loops and reverse one
  Split the hybgen regrid loops that correct for layers that are denser and
lighter than their targets, and work upward in the loop for the denser water to
avoid an instability in which a large portion of the water can be collected in a
single if the layers are much denser than their targets.  This will change
answers when Hybgen regridding is used, but the changes are small if the layers
are near their target densities.  Answers are unchanged if Hybgen regridding is
not being used.