+Add runtime parameter VELOCITY_REMAPPING_SCHEME
  Added a new runtime option to specify a different remapping scheme for
velocities than is used for tracers, as is done in Hycom.  By default,
velocities still the same scheme as tracers.  Also modified the interface to
remap_all_state_vars to take just the ALE_CS, which then provides the remapping
control structure that is appropriate for the tracers or velocities, rather than
also passing this in as an argument.  By default, all answers are bitwise
identical.