(*)+Replace hybgen_ppm_remap with hybgen_weno_remap
  Use the same ramapping code with both of the PPM-based hybgen remapping
schemes.  The WENO version was chosen for this because it seems to be a better
formulation, and is the more modern and commonly used of the two.  Note that
because the previous name is being reused, the code change in MOM_ALE might
appear to be the opposite of what it is actually.  This is mathematically
equivalent, but because there are differences between large numbers with both
schemes, this will change answers when the HYBGEN_PPM scheme is used for
remapping.