(*)Avoid hard-coded values in hybgen_column_unmix
  Use the equation of state in place of an ad-hoc hard-coded approximation to
limit the temperature and salinity changes in hybgen_column_unmix in the case
where the deepest layer is too light. Also eliminated the epsil argument to
hybgen_column_unmix, which is no longer necessary to prevent division by zero.
This change eliminates 5 hard-coded dimensional variables from the code.  This
will change answers in cases when USE_HYBGEN_UNMIX is True, but are bitwise
identical in other cases.