+(*)Use h_neglect for hybgen remapping and regridding
  Use the same h_neglect for hybgen remapping and regridding as is used for the
other remapping and regridding schemes, instead of using the larger value of
1e-6 m.  Also eliminated the runtime parameter HYBGEN_H_THIN_REMAP, and
eliminated the recently added h_thin arguments to remapping_core_h and
build_reconstructions_1d.  This will change answers in cases that use hybgen
regridding of one of the hybgen remapping schemes, but answers are bitwise
identical in other cases.