(*)Avoid negative thicknesses in mixed_layer_restrat
  Enforce a minimum thickness of 0.5*Angstrom in the mixed_layer_restrat
routines.  The streamfunctions in these routines attempt to limit the
thicknesses to exceed Angstrom, but they can be less than this due to roundoff.
The new limit prevents thicknesses from becoming negative when Angstrom is
set to 0, but should not change any answers for test cases with positive
values of Angstrom.  Also added some comments describing variables and their
units and simplified the OMP directives.  All answers in the existing
MOM6-examples regression suite are bitwise identical.