+Split MOM_debugging_attic out of MOM_debugging
  Created the new module MOM_debugging_attic to hold the vec_chksum, totalTandS
and totalStuff routines that are not being called from anywhere in the MOM6
code, but might be useful as temporary additions for future debugging exercises.
This changes where several publicly visible routines are found, and it
eliminates several get_param calls from debugging_init, so there are changes in
the MOM_parameter_doc.debugging files, but all answers are bitwise identical.