+(*)Fix USE_QG_LEITH_VISC = True reproducibility
  This commit includes a series of changes that get MOM6 to reproduce across
processor counts, layouts and restarts when USE_QG_LEITH_VISC = True.  These
changes include:

  - Expanding the halos to 2 in updates for h, tv%T and tv%S

  - Adding a thermo_var_ptrs and a timestep argument to to the interfaces to
    horizontal_viscosity and initialize_dyn_split_RK2

  - Adding the new publicly visible routine calc_QG_slopes

  - Adding the arguments slope_x and slope_y to calc_QG_Leith_viscosity and
    avoiding the use of the VarMix%slope_[xy] elements in that routine

  - Increasing some loop extents in calc_QG_Leith_viscosity

Note that several of the expansions of the halo sizes in updates or in some
of the calculations are not necessary if an extra halo update is included for
slope_x and slope_y after the call to calc_QG_slopes in horizontal_viscosity.
This extra halo update has also been included.  As it stands, this commit is a
working draft intended for discussion; once there is a collective decision about
how to handle some of the options here, there should be some clean-up before
this code is merged into the main branch of MOM6.  This commit includes several
changes to publicly visible interfaces, and it will change answers with
USE_QG_LEITH_VISC = True, but answers and output are bitwise identical in other
cases.