IO: netCDF reads changed to intent(inout)
This patch changes the intent of `values` from `intent(out)` to
`intent(inout)` in `get_field_nc` and `netCDF_read_data_2d`.
`read_netcdf_field` is left as `intent(out)`.

This is due to the ambiguous shape of `values` in these two functions,
which may represent either the compute or data domain.

Because only the compute domain of `values` is filled by `get_field_nc`,
it may contain residual values in its halos.  Previous use of
`intent(out)` would not necessarily preserve these values.  This change
explicitly directs the compiler to preserve them.