*Cancel Z_to_H factors in MOM_internal_tides.F90
  Cancelled out factors of GV%Z_to_H in MOM_internal_tides.F90 to simplify the
code and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq mode.
This involved changing the units of one internal variable in propagate_int_tide
and one element in the int_tide_CS type to use thickness units or their
inverse.  The variable I_rho0 was also eliminated and replaced with the
appropriate equivalent variable GV%RZ_to_H. Because GV%Z_to_H is an exact power
of 2 in Boussinesq mode, all answers are bitwise identical in that mode, but in
non-Boussinesq mode this conversion involves multiplication and division by
GV%Rho_0, so while all answers are mathematically equivalent, this change does
change answers at roundoff in non-Boussinesq mode.