*Cancel GV%Z_to_H in MOM_internal_tide_input.F90
  Cancelled out factors of GV%Z_to_H in MOM_internal_tide_input.F90 to simplify
the code and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq
mode.  This involved changing the units of 4 internal variables in
find_N2_bottom to use thickness units or related units.  In one place, GV%Rho0
was replaced with GV%H_to_RZ.  Because GV%Z_to_H is an exact power of 2 in
Boussinesq mode, all answers are bitwise identical in that mode, but in
non-Boussinesq mode this conversion involves multiplication and division by
GV%Rho_0, so while all answers are mathematically equivalent, this change does
change answers at roundoff in non-Boussinesq mode.