+*Cancel out Z_to_H factors in MOM_CVMix_KPP.F90
  Cancelled out factors of GV%Z_to_H in MOM_CVMix_KPP.F90 to simplify the code
and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq mode.  This
involved changing the units of 4 internal variables in KPP_calculate, 15
internal variables in KPP_compute_BLD, 5 internal variables in KPP_smooth_BLD
and 8 elements in the KPP_CS type to use thickness units or related units. The
previously incorrect description of the units the averaged velocities were also
fixed.

  This change also required the addition of a verticalGrid_type argument to
KPP_get_BLD, and corresponding changes to the calls to this routine from the
diabatic driver, neutral diffusion and lateral boundary diffusion code.

  Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.  There is also a new argument to a publicly visible
routine.