*Cancel out Z_to_H factors in MOM_barotropic.F90
  Cancelled out factors of GV%Z_to_H in MOM_barotropic.F90 to simplify the code
and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq mode.  This
involved changing the units of 4 variables in the barotropic_CS type, 3 internal
variables in btstep and an internal variable in barotropic_init to use thickness
units.  The rescaled internal variable mass_to_Z was also replaced with the
equivalent GV%RZ_to_H. Because GV%Z_to_H is an exact power of 2 in Boussinesq
mode, all answers are bitwise identical in that mode, but in non-Boussinesq mode
this conversion involves multiplication and division by GV%Rho_0, so while all
answers are mathematically equivalent, this change does change answers at
roundoff in non-Boussinesq mode.