*Cancel out Z_to_H factors in MOM_wave_speed.F90
  Cancelled out factors of GV%Z_to_H in MOM_wave_speed.F90 to simplify the code
and reduce the dependence on the value of GV%Rho_0 in non-Boussinesq mode.  This
involved changing the units of 21 internal variables in wave_speed and 16
internal variables in wave_speeds to use thickness units or other related
units.  Two internal variables used for the unit conversions were also renamed.
Because GV%Z_to_H is an exact power of 2 in Boussinesq mode, all answers are
bitwise identical in that mode, but in non-Boussinesq mode this conversion
involves multiplication and division by GV%Rho_0, so while all answers are
mathematically equivalent, this change does change answers at roundoff in
non-Boussinesq mode.