Removed h_neglect args from remapping_core_h()
The API for remaping_core_h() should not have h_neglect as an argument,
so converting it from optional to mandatory has moved things in the wrong
direction. The better way to get the *truly optional* parameter into the
remapping is as a parameter in the type. This commit implements that for
the remapping type. In the "big" refactor, the parameters are further moved
down to be associated with the actual reconstruction schemes.