*Non-Boussinesq code for RESET_INTXPA_INTEGRAL
  Added non-Boussinesq versions of the code that is exercised when
CORRECTION_INTXPA_5PT and RESET_INTXPA_INTEGRAL are set to true.  These options
use the same names as in their Boussinesq forms, even though the arrays that are
being adjusted are actually intx_za and inty_za.  Also changed the units of the
control structure variable that determines the surface displacement at which
surface mass-weighting in the pressure gradient force is considered to be active
from vertical distance units into thickness units.  As a part of the testing of
this commit, several checksums were added to the PressureForce_FV routines;
these are enabled when DEBUG = True.  The changes in this commit will change the
answers in non-Boussinesq cases when either of those options are in use, but are
bitwise identical when they are not, and all Boussinesq answers are bitwise
identical.