+Add post_tracer_integral_diagnostics
  Added the new routine post_tracer_integral_diagnostics to calculate vertically
integrated tracer diagnostics.  This includes properly setting the vertical
layer extents from layer thicknesses via a call to thickness_to_dz.  This new
routine is being called from within step_MOM(), immediately after another
closely related call to write out other tracer diagnostics.  In so doing, it
changes the thicknesses used for these diagnostics to be consistent with the
state of the tracers.

  This commit also sets the appropriate unit conversion factors in the
register_diag_field calls for three recently diagnostics of surface tracer
concentrations and the integrated tracers amounts in the full water column and
in the topmost 100 m.   All solutions are bitwise identical, but this will
correct the thicknesses used in the calculation of two groups of tracer
diagnostics and properly implements the dimensional rescaling of these
diagnostics when tracers (like temperature and salinity) when they are being
rescaled.