+Refactor stochastics for dimensional consistency
  Refactored MOM_stochastics for dimensional and rotational consistency.  This
involved correcting the dimensions of a hard-coded area, adding missing
conversion factors to the register_diag_field calls for 6 stochastics
diagnostics, and adding descriptions of the units in comments describing 7
variables in MOM_stochastics.  This commit also adds the new runtime parameter
STOCHASTICS_ANSWER_DATE that can be set to a value of 20250701 or higher to use
rotationally symmetric expressions.  The stochastic physics package itself is
external to MOM6 and works in unscaled mks units, so fields passed to it (such
as the timestep) need to be unscaled back to mks units.   As a part of this
change, new unit_scale_type arguments were added to stochastics_init and
apply_skeb.  By default the answers should be bitwise identical in cases without
dimensional consistency testing, but the diagnostic conversion factors should
correct the problems with dimensional rescaling.