+Add DEBUG_OBCS and OBC_DEBUGGING_TESTS
  Added the new runtime parameter DEBUG_OBCS that triggers verbose diagnostic
output about the contents of the open boundary condition type via calls to the
new routines write_OBC_info() and chksum_OBC_segments(), along with the new
runtime parameter NK_OBC_DEBUG to regulate the volume of diagnostic output.  It
also renames the previous runtime parameter DEBUG_OBC to OBC_DEBUGGING_TESTS to
trigger the code that resets the normal velocities at OBC segments to zero and
the thicknesses and tangential velocities behind OBC segments to silly values
for testing, with DEBUG_OBC retained for now as the old name.  This latter
change follows the pattern elsewhere in the code in which other DEBUG parameters
do not actually change any values.  This commit also comments out a fatal error
message if open boundary conditions are applied with rotations of 180 or 270
degrees.  All solutions in cases that worked previously are bitwise identical,
but there are new runtime parameters.