(*)Copy tracer registry in rotate_OBC_segment_data
  Modified rotate_OBC_segment_data to copy over information in the tr_Reg
elements of the segment when they are in use.  Prior to this change, there would
be a segmentation fault if the segment tracer registry was in use and the index
rotation was being enabled (even if the grid was not actually being rotated).
All answers are bitwise identical in cases that worked previously, but some
cases are now able to use rotated indices that were not previously able to do
so.