Add ALE_COORDINATE_CONFIG HYBRID_3D and HYBRID_MAP
Add two options supporting 3D HYCOM1 TARGET_DENSITIES and ALE_RESOLUTION.

The fully general interface is ALE_COORDINATE_CONFIG=HYBRID_3D which
is like HYBRID except that the netCDF variables (e.g. sigma2 and dz)
are 3D fields (x-y-z).  Like HYBRID, the 2nd variable (e.g. dz) can be
replaced by a FNC1 string that sets a dz profile that is used everywhere.

The typical use of 3D HYCOM1 fields is to have different TARGET_DENSITIES
in semi-enclosed seas for layers that are always below its sill depth.
For example, in the OM4 75-layer setup layers 71 amd 72 are only active
in the Arctic and layers 74 and 75 are only active in the Mediterranean.
So the number of layers can be reduced to 71 by specifying different
deep targets in these semi-emclosed seas.  In this case ALE_RESOLUTION
would be the same everywhere, but the resulting 71 layer configuration
can be further improved by using shallow ALE_RESOLUTION in the northern
hemisphere and the standard, deep, dz's in the southern hemisphere.
This is safe, because deep layers are never in fixed coordinates near
the equator.

To simplify this typical use case, ALE_COORDINATE_CONFIG=HYBRID_MAP
requires three netCDF variables (e.g. map,sigma2,dz) with the last
two being 2-D z-index fields containing a small number of profiles
and the first a 2-D x-y field of index values indicating which profile
to use at each location.  Like HYBRID and HYBRID_3D, the 3rd variable
(e.g. dz) can be replaced by a FNC1 string that sets a dz profile
that is used everywhere.

Answers are not changes unless ALE_COORDINATE_CONFIG is set to HYBRID_3D
or HYBRID_MAP.