Move calls out of MOM_initialize_OBCs
  Moved various calls out of MOM_initialize_OBCs to simplify the OBC
initialization code and the amount of rotation-specific OBC code that is needed.
The call to initialize_segment_data is moved into initialize_MOM adjacent to
other OBC configuration code that depends on the vertical grid.  The call to
update_OBC_segment_data now occurs later in initialize_MOM and is now shared
between the cases where there is and is not grid rotation in use.  There is no
longer any need to retain set_tracer_data, so it has been removed.  In addition,
changes were made to supercritical_set_OBC_data to support rotated grids.  All
answers are bitwise identical.