Adjust rate of adding mass source to BT solver
The rate at which eta_src was added to eta within the barotropic
solver allowed the instantaneous eta to ground-out and drop below
the bottom during the extend steps over which the filtered state
is calculated. eta_src was previously adjusted so that it could
at worst ground out over the course of the baroclinic step, but
did not allow for the extra time needed when filtering.

This change was found to be helpful when debugging problems in
vanished layers under a grounded thermodynamic ice shelf.