Enhanced user-friendliness: OBC configuration consistency check included (#936)
* OBC configuration consistency check included

In the subroutine parse_segment_str, a check is now executed to ensure the number of OBC types for each segment is not above 8. The model now stops if more than 8 segments are configured in MOM_input.

* modification of OBC string consistency check

the maximum number of OBC types is now derived from the length of action_str