Adds limiter for time-integrated BT transports
The BT solver regularly spits out warnings about the SSH dropping below
the ocean bottom, particularly frequently for extreme scenarios such as
the vicinity of the ground line in an ice shelf cavity. These changes
apply a limiter on the time-integrated transport within the BT solver
that based on the initial volume adjusted for any sinks.

- Added run-time flag BT_LIMIT_INTEGRAL_TRANSPORT that turns on
  calculation of the initial and projected available volume.
  By default, this is off and the default available volume is huge.
- Implemented a limiter without `if`s for efficiency.
- Added a FATAL test to check that the initial conditions do not already
  have negative total thickness.
- Flipped the WARNING about negative total thickness to be FATAL when
  using the limiter, since if the limiter is not robust then we've done
  something wrong again.
- Extended check on number of steps for BT solver, wrt filtering steps
  - Changed an ==0 to <=0, since we encountered a weird situation where
    it happened (unknown cause)