+Separate OBC_for_remap and OBC_for_bug arguments
  Add separate OBC_for_remap and OBC_for_bug arguments to MOM_initialize_state()
to clarify the distinct roles of the two arguments, to make it clear that one
can not be used without rotation, and to prepare for the eventual obsoleting of
OBC_RESERVOIR_INIT_BUG.  This commit also simplifies the logic setting ntstep
when do_thermo is false. All answers are bitwise identical, but there is a new
optional argument to MOM_initialize_state and another argument has been renamed.