+Add chksum_OBC_segment_data
  Added the new subroutine chksum_OBC_segment_data() from the contents of the
loop over the segments in chksum_OBC_segments, as this can be useful when called
separately for debugging some of the constructs that will be coming in
subsequent commits.  Also added rotate_OBC_segment_values_needed() to copy over
the values_needed and _index variables between segments, and to avoid the need
for duplicate code blocks in rotate_OBC_segment_data() and rotate_OBC_config().
There is also some minor revisions to rename segnam to segname for greater
clarity.  This commit also makes greater use of the I0 format that was
introduced with Fortran 95 to simplify or shorten some error messages.  All
answers are bitwise identical but there are two new internal subroutines in the
MOM_open_boundary module.