+Call initialize_segment_data with rotated OBCs
  Call initialize_segment_data() with the rotated OBC type.  This includes the
addition of a new turns arguments to initialize_segment_data() and the
elimination of the ocean_grid_type argument to this routine.  This includes the
addition of the new function rotated_field_name() to swap the names of u- and v-
velocity fields when the grid is rotated, avoiding the duplication of code.
Internally in initialize_segment_data(), num_fields was renamed to reflect the
fact that it is actually the number of tracer or other fields that are
initialized via the OBC manifest string (which does not include generic OBGC
tracers), and not the total number of fields on a segment.  All answers are
bitwise identical, but there are changes to the arguments of the publicly
visible routine initialize_segment_data().