+Eliminate rotate_OBC_segment_fields
  Eliminated the now unused subroutines rotate_OBC_segment_fields(),
rotate_OBC_segment_tracer_registry(), rotate_OBC_segment_data() and
rotate_OBC_segment_tracer_data(), along with a commented out block of code in
rotate_OBC_config().  Also eliminated the duplicate call to
register_temp_salt_segments() on the unrotated OBC type, as this is no longer
used to populate the rotated OBC type.  Also corrected the do_not_log argument
for DT_OBG_SEG_UPDATE_OBGC so that it will get logged when it is being used, and
moved the call to open_boundary_setup_vert() later in the code so that it is in
the same block as other OBC code.  All answers are bitwise identical but there
are two fewer publicly visible OBC-related interfaces and there will be new
entries in some MOM_parameter_doc files.