Fix calculation of CAv_Stokes diagnostic
  Corrected a horizontal indexing bug in the calculation of the CAv_Stokes
diagnostic, making it rotationally consistent and consistent with the
calculation of CAu_Stokes.  This bug has been there since the CAv_Stokes
diagnostic was originally added.  The loop range over which qS is calculated was
also reduced to the range over which it is used.  All solutions are bitwise
identical, but this commit does change the values of a (perhaps infrequently
used) diagnostic.